requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.12.0
django-filter==25.1
pymysql
pytest==8.4.1
pytest-django==4.11.1
mysqlclient
drf-spectacular==0.28.0
Pillow==11.3.0
coverage==7.10.3


django-environ
django-cors-headers==4.7.0
django-filter==25.1
drf_yasg
#mysqlclient==2.2.0
pymysql
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0


factory-boy


upgrade pip